refactor(video-player): isolate progress observation
